+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-- Added `Gacela::get('rootDir')`
+- Added `Gacela::rootDir()`
 
 ### 1.3.0
 ### 2023-05-08